Working on getting C to call Python function updateSoundArray.
It was going well and I managed to run the program except changing instruments had not been implemented yet. Since doing this the program now compiles and produces an executable that doesn't run. The file manager thinks it's a text file and when run from terminal it produces a Segmentation Fault before even running main(). Possible causes include the way we link in the Makefile and the uses of Py_* functions.